requirements.txt
PyYAML==3.12
boto3>=1.6.16
botocore>=1.9.16
azure-mgmt-authorization
azure-mgmt-network
azure-cli-core
